move DonutSlice into its own file /cc @emplums
