+++ src/DonutChart.js
@@ -1,19 +1,7 @@
 import React from 'react'
+import DonutSlice from './DonutSlice'
 import PropTypes from 'prop-types'
 import {arc as Arc, pie as Pie} from 'd3-shape'
-import theme from './theme'
-
-const {colors} = theme
-
-const fillForState = {
-  'error': colors.red[5],
-  'queued': colors.yellow[7],
-  'pending': colors.yellow[7],
-  'failure': colors.red[5],
-  'success': colors.green[5]
-}
-
-const DEFAULT_FILL = colors.gray[4]
 
 function mapData(data) {
   return Object.keys(data)
@@ -35,35 +23,24 @@ const DonutChart = props => {
   const pie = Pie()
     .value(child => child.props.value)
 
+  // coerce the children into an array
   const childList = React.Children.map(children, d => d)
-  const arcData = childList.length ? pie(childList) : []
+  const arcData = pie(childList)
   const arc = Arc()
     .innerRadius(innerRadius)
     .outerRadius(radius)
 
-  const arcs = childList.map((child, i) => {
-    const {
-      state,
-      fill = fillForState[state] || DEFAULT_FILL
-    } = child.props
-    return React.cloneElement(child, {
-      d: arc(arcData[i]),
-      fill,
-      key: i
-    })
+  const slices = childList.map((child, i) => {
+    return <DonutSlice {...child.props} d={arc(arcData[i])} key={i} />
   })
 
   return (
     <svg width={size} height={size}>
       <g transform={`translate(${radius},${radius})`}>
-        {arcs}
+        {slices}
       </g>
     </svg>
   )
-}
-
-const DonutSlice = ({children, d, fill}) => {
-  return <path d={d} fill={fill}>{children}</path>
 }
 
 // see: <https://github.com/facebook/react/issues/2979>
@@ -80,12 +57,4 @@ DonutChart.propTypes = {
   size: PropTypes.number
 }
 
-DonutSlice.propTypes = {
-  d: PropTypes.string,
-  fill: PropTypes.string,
-  state: PropTypes.oneOf(Object.keys(fillForState)),
-  value: PropTypes.number
-}
-
 export default DonutChart
-export {DonutSlice}
