+++ src/index.js
@@ -18,7 +18,9 @@ export { default as ButtonOutline } from './ButtonOutline'
 export { default as ButtonLink } from './ButtonLink'
 
 export { default as Details } from './Details'
-export { default as DonutChart, DonutSlice} from './DonutChart'
+
+export { default as DonutChart} from './DonutChart'
+export { default as DonutSlice} from './DonutSlice'
 
 export { default as Heading } from './Heading'
 export { default as Label } from './Label'
